Client C program file is compiled using gcc and has no command line parameters. e.g :- ./client Server C program file is compiled using gcc and has no command line parameters. e.g :- ./server
Program 2:
The program 2 uses fork and exec family command to execute the ls with
parameter passed to it by the client. Then the output of the ls is redirected using
output redirection to the pipe. And from the pipe the output of ls is read and sent to
the client.