Server C program file is compiled using gcc and takes number of cylinders, number of sectors and name of file for disk as command line parameters. e.g. :- ./server 4 2 disk.txt When the client connects with the server it asks user to enter the valid commands to be executed on the basic file system. The server receives those commands and execute them on basic file system and send the particular output to the client user.
In program 3, a struct is used to store the information about the disk such as disk
size, number of cylinders, number of sectors, pointer to the disk. We used that
struct to make the functions/ commands to work. In reading and writing we
calculate the offset using the cylinder number and sector number in-order to
write/read data from the exact block of the memory.