Updated Server.c Code & Makefile
In program 4, a struct is used to store the information about the disk such as disk
size, number of cylinders, number of sectors, pointer to the disk. A struct for file
system is maintained in which Bitmap is used to keep track of the free blocks and
number of files in the system. And a struct to maintain information about each file
in the FAT system such as block number allocated to the particular file, its name,
and size of file. The file system struct is used to make the function/commands
work in sync with the disk. Whenever a block is allocated to the file the bit of that
block in the Bitmap is set and when it is freed by the file, the block number is freed
in the Bitmap and made available for writing. A bitmap is a char * array so that
each char can accommodate 8 blocks.